+++ static/project-template/src/components/hello-world/hello-world.component.tsx
@@ -1,0 +1,9 @@
+import styles from "./hello-world.module.scss";
+
+export function HelloWorld() {
+  return (
+    <div id="hello-world" className={styles["text"]}>
+      hello world!!!
+    </div>
+  );
+}