+++ static/project-template/src/components/hello-world/hello-world.script.tsx
@@ -1,0 +1,5 @@
+window.addEventListener("load", () => {
+  const helloWorldDiv = document.querySelector<HTMLElement>("#hello-world");
+  if (helloWorldDiv === null) return;
+  helloWorldDiv.style.border = "2px solid #00f";
+});