+++ static/project-template/src/types/torytis-variable.d.ts
@@ -4,7 +4,5 @@
  * 주의! 본 파일에 커스텀 내용을 추가할 경우 해당 내용은 torytis 특정 명령어 구동시에 유실 될 수 있습니다.
  */
 declare namespace JSX {
-  interface IntrinsicElements {
-//THIS_IS_REPLACE_SPOT//
-  }
+  interface IntrinsicElements {}
 }